requirements_dev.txt
Django==1.5.1
Fabric==1.6.1
South==0.8.1
distribute==0.6.24
djangorestframework==2.3.5
paramiko==1.10.1
pycrypto==2.6
wsgiref==0.1.2
